requirements.txt
streamlit
opencv-python
whisper
sentence-transformers
faiss-cpu
numpy
ffmpeg-python
openai